Update bind.js
I have implemented a way of differentiating between a hard coded channel id that is valid and a not valid every other channel on the server (.env file has the channel id stored there). Now, I will attempt to change the message event to validate that the correct channel id is being used or not.