Update messsage.js
Atm, I have just hard coded the channel id in the .env so that it now checks before a command is executed that it is on the right channel. I would like to implement this feature using a bind command but without a database to get the channel id from and without the ability to set a global flag in the .env file to state whether a channel bind is true or false, I don't believe I can go further with implementing this.