Update bind.js
Updated the .env file to have an empty variable to store the channel id that gets binded. In the Bind command, it will now check that if it is empty, bind the channel, if it is already set to that channel, send a message saying so. Otherwise, it will decline the bind command.